fix type

+++ urdf2webots/writeProto.py
@@ -285,7 +285,7 @@ def URDFShape(proto, link, level, normal=False):
                     if len(value) == 3:
                         proto.write('%d %d %d -1 ' % (value[0], value[1], value[2]))
             elif isinstance(visualNode.geometry.trimesh.coordIndex[0], np.int32):
-                for i in range(len(visualNode.geometry.trimesh.coordIndex) / 3):
+                for i in range(int(len(visualNode.geometry.trimesh.coordIndex) / 3)):
                     proto.write('%d %d %d -1 ' % (visualNode.geometry.trimesh.coordIndex[3 * i + 0], visualNode.geometry.trimesh.coordIndex[3 * i + 1], visualNode.geometry.trimesh.coordIndex[3 * i + 2]))
             else:
                 print('Unsupported "%s" coordinate type' % type(visualNode.geometry.trimesh.coordIndex[0]))
